@@ -2,36 +2,18 @@
   "name": "client",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:5000/",
   "dependencies": {
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.1",
-    "@mui/styled-engine-sc": "^5.4.2",
-    "@shopify/polaris": "^7.5.0",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.24.0",
     "dat.gui": "^0.7.9",
-    "firebase": "^9.6.9",
-    "gsap": "^3.5.1",
-    "react": "^17.0.2",
-    "react-beautiful-dnd": "^13.0.0",
-    "react-color": "^2.19.3",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "5.0.0",
-    "react-simple-image-slider": "^2.3.0",
-    "redux": "^4.1.2",
-    "styled-components": "^5.3.3",
-    "three": "^0.138.3",
-    "web-vitals": "^2.1.3"
+    "three": "^0.124.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "publish": "npm run build && serve -s build",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
