@@ -1,20 +1,21 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import Index from "./App";
-import reportWebVitals from "./reportWebVitals";
-import { AppProvider } from "@shopify/polaris";
-import { BrowserRouter } from "react-router-dom";
-
-import enTranslations from "@shopify/polaris/locales/en.json";
-import "@shopify/polaris/build/esm/styles.css";
-import "./css/font.css";
-ReactDOM.render(
-  <BrowserRouter basename="/">
-    <AppProvider i18n={enTranslations}>
-      <Index />
-    </AppProvider>
-  </BrowserRouter>,
-  document.getElementById("root")
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
+import reportWebVitals from './reportWebVitals';
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
 );
 
+// ReactDOM.render(
+//     <App />,
+//   document.getElementById('root')
+// );
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
